refactor(storage): rename 'abstract.py' to 'abc.py'
