+++ src/mydb/storage/abc.py
@@ -17,6 +17,10 @@ class StorageEngine(ABC):
 
 class Index(ABC):
     @abstractmethod
+    def has(self, key: bytes, /) -> bool:
+        raise NotImplementedError
+
+    @abstractmethod
     def set(self, key: bytes, offset: int, /) -> None:
         raise NotImplementedError
